Implement `Gcd<Uint>` for `Int`
